Adding Back Alleles and Segments
After talking with Benedict, it makes sense to me what these are for and why we need them. We need to be able to make calls on portions of variants, so we can call the bit of a variant between two other variants on it, for example.
And if we can do that, we really ought to be able to call several of these Segments at once as one Allele (or so Benedict says). And then we ought to be able to call both ends of, say, an inversion together as one Allele, without reading through the whole thing.

And then we need the Scaffolds back to read through individual Alleles and not multi-Allele-at-once Calls. Abd they need to be associated with call sets and variant sets.