Adding documentation
Adding in Doxygen capabilities for generating UML.

Adding UML diagram.

Specifying how the UML diagram can be generated.

Automating Avrodoc build with a script.

Adding proper Beacon stuff to the UML

Updating UML to drop AlleleResource

Adding a Graph Mode FAQ

It would be good to have the answers to people's questions about graph mode
all in one place.

Moving and renaming documentation

All the extra Markdowns should go in doc/, and should not have spaces
in the filenames.

Adding an SVG of the UML to the repo.